requirements.txt
Flask
flask-cors
tensorflow==2.13.0
gdown
Pillow
numpy
gunicorn
